fixing issues in pull request per Sean's direction

+++ Source/Plugins/SurfaceMeshing/SurfaceMeshingFilters/FindGBCD.cpp
@@ -648,15 +648,15 @@ void FindGBCD::sizeGBCD(int faceChunkSize, int numMisoReps)
   //m_GBCDlimits[9] = cosf(0.0);
 
   //Greg's Ranges
-  m_GbcdLimits[0] = 0.0;
-  m_GbcdLimits[1] = 0.0;
-  m_GbcdLimits[2] = 0.0;
-  m_GbcdLimits[3] = 0.0;
-  m_GbcdLimits[4] = 0.0;
+  m_GbcdLimits[0] = 0.0f;
+  m_GbcdLimits[1] = 0.0f;
+  m_GbcdLimits[2] = 0.0f;
+  m_GbcdLimits[3] = 0.0f;
+  m_GbcdLimits[4] = 0.0f;
   m_GbcdLimits[5] = DREAM3D::Constants::k_PiOver2;
-  m_GbcdLimits[6] = 1.0;
+  m_GbcdLimits[6] = 1.0f;
   m_GbcdLimits[7] = DREAM3D::Constants::k_PiOver2;
-  m_GbcdLimits[8] = 1.0;
+  m_GbcdLimits[8] = 1.0f;
   m_GbcdLimits[9] = DREAM3D::Constants::k_2Pi;
 
   float binsize = m_GBCDRes * DREAM3D::Constants::k_PiOver180;
@@ -677,10 +677,10 @@ void FindGBCD::sizeGBCD(int faceChunkSize, int numMisoReps)
   float totalNormalBins = m_GbcdSizes[3] * m_GbcdSizes[4];
   m_GbcdSizes[3] = int(sqrt(totalNormalBins) + 0.5);
   m_GbcdSizes[4] = int(sqrt(totalNormalBins) + 0.5);
-  m_GbcdLimits[3] = -sqrt(DREAM3D::Constants::k_PiOver2);
-  m_GbcdLimits[4] = -sqrt(DREAM3D::Constants::k_PiOver2);
-  m_GbcdLimits[8] = sqrt(DREAM3D::Constants::k_PiOver2);
-  m_GbcdLimits[9] = sqrt(DREAM3D::Constants::k_PiOver2);
+  m_GbcdLimits[3] = -sqrtf(DREAM3D::Constants::k_PiOver2);
+  m_GbcdLimits[4] = -sqrtf(DREAM3D::Constants::k_PiOver2);
+  m_GbcdLimits[8] = sqrtf(DREAM3D::Constants::k_PiOver2);
+  m_GbcdLimits[9] = sqrtf(DREAM3D::Constants::k_PiOver2);
   m_GbcdDeltas[3] = (m_GbcdLimits[8] - m_GbcdLimits[3]) / float(m_GbcdSizes[3]);
   m_GbcdDeltas[4] = (m_GbcdLimits[9] - m_GbcdLimits[4]) / float(m_GbcdSizes[4]);
 }
